+++ src/openassetio-core/include/openassetio/TraitsData.hpp
@@ -48,8 +48,8 @@ OPENASSETIO_DECLARE_PTR(TraitsData)
  * "well-known-strings") is inherently unstable. Instances of this class
  * should generally be wrapped in one of the specialized @ref
  * openassetio.SpecificationBase "Specification" or @ref
- * openassetio.Trait "Trait" derived "views" at runtime by a @ref host
- * or @ref manager to ensure consistent access to the correct keys.
+ * openassetio.TraitBase "Trait" derived "views" at runtime by a @ref
+ * host or @ref manager to ensure consistent access to the correct keys.
  *
  * Trait @ref trait::property::Key "property keys" are always strings.
  * Property values are strings, integers, floating point, or booleans.